droits...

+++ src/models/PageDAO.class.php
@@ -5,36 +5,32 @@ require_once(MODELS_INC."Page.class.php");
 
 class PageDAO
 {
-
 	public static function getAll()
 	{
-		try{
+		try {
 			$bdd=connect();
-			$req=$bdd->query("SELECT `idPage`, `libelle` FROM `page` ORDER BY libelle");
+			$req=$bdd->query("SELECT * FROM `page` ORDER BY libelle");
 			$lst=$req->fetchAll();
 			$lstObj=array();
-			foreach
-			($lst as $page)
+			foreach ($lst as $page)
 			{
 				$lstObj[]=new Page($page['idPage'], $page['libelle']);
 			}
 			return $lstObj;
-		}catch(PDOException $e)
-		{
+		} catch(PDOException $e) {
 			die('error get all page '.$e->getMessage().'<br>');
 		}
 	}
 
 	public static function getById($id)
 	{
-		try{
+		try {
 			$bdd=connect();
-			$req=$bdd->prepare("SELECT `idPage`, `libelle` FROM `page` WHERE idPage=?");
+			$req=$bdd->prepare("SELECT * FROM `page` WHERE idPage=?");
 			$req->execute(array($id));
 			$page=$req->fetch();
 			return new Page($page['idPage'], $page['libelle']);
-		}catch(PDOException $e)
-		{
+		} catch(PDOException $e) {
 			die('error get id page '.$e->getMessage().'<br>');
 		}
 
@@ -42,14 +38,13 @@ class PageDAO
 
     public static function getByLibelle($lib)
 	{
-		try{
+		try {
 			$bdd=connect();
-			$req=$bdd->prepare("SELECT `idPage`, `libelle` FROM `page` WHERE libelle=?");
+			$req=$bdd->prepare("SELECT * FROM `page` WHERE libelle=?");
 			$req->execute(array($lib));
 			$page=$req->fetch();
 			return new Page($page['idPage'], $page['libelle']);
-		}catch(PDOException $e)
-		{
+		} catch(PDOException $e) {
 			die('error get libelle page '.$e->getMessage().'<br>');
 		}
 
@@ -57,8 +52,7 @@ class PageDAO
 
 	public static function create(&$page)
 	{
-		if
-		(gettype($page)=="Page")
+		if (gettype($page)=="Page")
 		{
 			try{
 				$bdd->connect();
@@ -66,53 +60,44 @@ class PageDAO
 				$req->execute(array($page->getLibelle()));
 				$page->setId($bdd->LastInsertId());
                 return $page->getId();
-			}catch(PDOException $e)
-			{
+			} catch(PDOException $e) {
 				die('error create page '.$e->getMessage().'<br>');
 			}
-		}else
-		{
+		} else {
 			die('paramètre de type page requis');
 		}
 	}
 
 	public static function update($page)
 	{
-		if
-		(gettype($page)=="Page")
+		if (gettype($page)=="Page")
 		{
-			try{
+			try {
 				$bdd->connect();
 				$req=$bdd->prepare("UPDATE `page` SET `libelle`=? WHERE `idPage`=?");
 				$req->execute(array($page->getLibelle(), $page->getId()));
-			}catch(PDOException $e)
-			{
+			} catch(PDOException $e) {
 				die('error update page '.$e->getMessage().'<br>');
 			}
-		}else
-		{
+		} else {
 			die('paramètre de type page requis');
 		}
 	}
 
 	public static function delete($page)
 	{
-		if
-		(gettype($page)=="Page")
+		if (gettype($page)=="Page")
 		{
-			try{
+			try {
 				$bdd->connect();
 				$req=$bdd->prepare("DELETE FROM `page` WHERE `idPage`=?");
 				$req->execute(array($page->getId()));
-			}catch(PDOException $e)
-			{
+			} catch(PDOException $e) {
 				die('error delete page '.$e->getMessage().'<br>');
 			}
-		}else
-		{
+		} else {
 			die('paramètre de type page requis');
 		}
 	}
-
 }
 ?>
